# backend/docs/swagger.json
{
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "HTTP API онлайн-сервиса для IT-школы.",
        "title": "skadi API",
        "contact": {},
        "version": "1.0.0"
    },
    "host": "127.0.0.1:8000",
    "basePath": "/api/v1",
    "paths": {
...
        "/solution/{id}": {
            "get": {
                "security": [
                    {
                        "JWTAccess": []
                    }
                ],
                "description": "Получение всех данных о решении задания с полной инфой о задании и преподе (ID и полное имя), а также со списком учеников, которые тоже выполняют это задание.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "task"
                ],
                "summary": "Получение решения задания по id [Преподаватель и ученик].",
                "operationId": "solution-read",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID решения задания",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/v1.solutionOut"
                        }
                    },
                    "401": {
                        "description": "неверный токен (пустой, истекший или неверный формат)"
                    },
                    "403": {
                        "description": "доступ запрещён"
                    },
                    "404": {
                        "description": "решение задания не найдено"
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "JWTAccess": []
                    }
                ],
                "description": "Удаление решения (не задания целиком) по его id.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "task"
                ],
                "summary": "Удаление решения по id. [Только преподаватель]",
                "operationId": "solution-delete",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID решения",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "401": {
                        "description": "неверный токен (пустой, истекший или неверный формат)"
                    },
                    "403": {
                        "description": "доступ запрещён"
                    }
                }
            }
        },
        "/student/solution": {
            "get": {
                "security": [
                    {
                        "JWTAccess": []
                    }
                ],
                "description": "Получение списка решений конкретного ученика.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "task"
                ],
                "summary": "Получение списка решений [Только ученик].",
                "operationId": "student-solution-list",
                "parameters": [
                    {
                        "type": "boolean",
                        "example": true,
                        "description": "filter for checked solutions if true",
                        "name": "archived",
                        "in": "query"
                    },
                    {
                        "minimum": 1,
                        "type": "integer",
                        "example": 1,
                        "description": "page pagination param",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "minimum": 1,
                        "type": "integer",
                        "default": 10,
                        "example": 5,
                        "description": "per page pagination param",
                        "name": "per-page",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/v1.listSolutionOut"
                        }
                    },
                    "401": {
                        "description": "неверный токен (пустой, истекший или неверный формат)"
                    }
                }
            }
        },
        "/student/solution/{id}": {
            "patch": {
                "security": [
                    {
                        "JWTAccess": []
                    }
                ],
                "description": "Частичное обновление решения (только переданные поля: статус, кроме \"проверено\", ответ, файл ответа) по его id.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "task"
                ],
                "summary": "Обновление решения. [Только ученик]",
                "operationId": "student-solution-update",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID решения",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "updateSolutionBody",
                        "name": "updateSolutionBody",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/v1.updateSolutionBody"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.Solution"
                        }
                    },
                    "400": {
                        "description": "неверный статус"
                    },
                    "401": {
                        "description": "неверный токен (пустой, истекший или неверный формат)"
                    },
                    "403": {
                        "description": "доступ запрещён"
                    },
                    "404": {
                        "description": "решение не найдено"
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "entity.Contact": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                }
            }
        },
        "entity.Pagination": {
            "type": "object",
            "required": [
                "page",
                "pages",
                "per_page",
                "total"
            ],
            "properties": {
                "page": {
                    "description": "current page",
                    "type": "integer",
                    "example": 1
                },
                "pages": {
                    "description": "total pages amount",
                    "type": "integer",
                    "example": 7
                },
                "per_page": {
                    "description": "per page objects",
                    "type": "integer",
                    "default": 10,
                    "example": 5
                },
                "total": {
                    "description": "total objects amount",
                    "type": "integer",
                    "example": 32
                }
            }
        },
        "entity.Profile": {
            "type": "object",
            "required": [
                "fullname"
            ],
            "properties": {
                "address": {
                    "description": "user address",
                    "type": "string"
                },
                "contact": {
                    "$ref": "#/definitions/entity.Contact"
                },
                "extra": {
                    "description": "extra data (for admin only)",
                    "type": "string"
                },
                "fullname": {
                    "description": "user full name",
                    "type": "string"
                },
                "id": {
                    "description": "profile id (equals to user id)",
                    "type": "integer"
                },
                "parent_contact": {
                    "$ref": "#/definitions/entity.Contact"
                }
            }
        },
        "entity.Solution": {
            "type": "object",
            "required": [
                "id",
                "status",
                "task"
            ],
            "properties": {
                "answer": {
                    "description": "solution text answer",
                    "type": "string"
                },
                "grade": {
                    "description": "solution grade",
                    "type": "string"
                },
                "id": {
                    "description": "solution id",
                    "type": "integer"
                },
                "status": {
                    "description": "status object",
                    "allOf": [
                        {
                            "$ref": "#/definitions/entity.Status"
                        }
                    ]
                },
                "student": {
                    "description": "student object",
                    "allOf": [
                        {
                            "$ref": "#/definitions/entity.Profile"
                        }
                    ]
                },
                "task": {
                    "description": "task object",
                    "allOf": [
                        {
                            "$ref": "#/definitions/entity.Task"
                        }
                    ]
                },
                "updated_at": {
                    "description": "last-update datetime of solution",
                    "type": "string"
                }
            }
        },
        "entity.Status": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "id": {
                    "description": "status id",
                    "type": "integer"
                },
                "name": {
                    "description": "status name",
                    "type": "string"
                }
            }
        },
        "entity.Task": {
            "type": "object",
            "required": [
                "id",
                "title"
            ],
            "properties": {
                "description": {
                    "description": "task description",
                    "type": "string"
                },
                "id": {
                    "description": "task id",
                    "type": "integer"
                },
                "teacher": {
                    "description": "teacher object",
                    "allOf": [
                        {
                            "$ref": "#/definitions/entity.Profile"
                        }
                    ]
                },
                "title": {
                    "description": "task title",
                    "type": "string"
                }
            }
        },
...
        "v1.listSolutionOut": {
            "description": "listSolutionOut represents a solution list data.",
            "type": "object",
            "required": [
                "data"
            ],
            "properties": {
                "data": {
                    "description": "solutions list",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entity.Solution"
                    }
                },
                "pagination": {
                    "description": "pagination params",
                    "allOf": [
                        {
                            "$ref": "#/definitions/entity.Pagination"
                        }
                    ]
                }
            }
        },
...
        "v1.solutionOut": {
            "description": "solutionOut represents a solution data with students (solving the same task).",
            "type": "object",
            "required": [
                "solution"
            ],
            "properties": {
                "other_students": {
                    "description": "other students solving the same task",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entity.Profile"
                    }
                },
                "solution": {
                    "description": "solution object",
                    "allOf": [
                        {
                            "$ref": "#/definitions/entity.Solution"
                        }
                    ]
                }
            }
        },
...
        "v1.updateSolutionBody": {
            "description": "updateSolutionBody represents a data with optional body to update solution.",
            "type": "object",
            "properties": {
                "answer": {
                    "description": "new answer",
                    "type": "string",
                    "example": "ООП - это объектно-ориентированное программирование"
                },
                "grade": {
                    "description": "new grade",
                    "type": "string",
                    "maxLength": 50,
                    "example": "5+"
                },
                "status_id": {
                    "description": "new status ID",
                    "type": "integer",
                    "example": 2
                }
            }
        },
...
    },
    "securityDefinitions": {
        "JWTAccess": {
            "description": "Access JWT-token. Cookie will automatic add after auth is done.",
            "type": "apiKey",
            "name": "access",
            "in": "cookie"
        },
        "JWTRefresh": {
            "description": "Refresh JWT-token. Cookie will automatic add after auth is done.",
            "type": "apiKey",
            "name": "refresh",
            "in": "cookie"
        }
    }
}